Makefile: Some possible fixes to build.
al5poly: Fix to prevent unnecessary rebuilds due to objdir prerequisite.

squirrel: When invoking make(1) from the top-level directory of the repo
only the deps/libal5poly submodule would be built and then it would stop
there even though it succeeded. I think that this was an attempt to fix
that, and maybe it worked. If this did indeed fix it I can't explain why
it worked before now though (or why I only noticed it not working now?).
Anyway, moving on? This change was made before the change to al5poly (I
think).