!!wip Getting a circle on the screen finally.
The culprit appears to be that al_map_rgba() is failing within the
library, but working fine directly... For now I have worked around it,
but I have asked the community for explanations.

This is a dirty commit with several days of hacks included. Sorry. Will
try to fix later when there's time. Haha. Yeah, right.